feature: 流水线记录失败插件至var--支持内嵌原子
