+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/PipelineTaskService.kt
@@ -30,15 +30,20 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.util.JsonUtil
+import com.tencent.devops.common.pipeline.Model
 import com.tencent.devops.common.pipeline.pojo.element.ElementAdditionalOptions
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.log.utils.LogUtils
 import com.tencent.devops.process.dao.PipelineTaskDao
 import com.tencent.devops.process.engine.control.ControlUtils
 import com.tencent.devops.process.engine.dao.PipelineModelTaskDao
+import com.tencent.devops.process.engine.pojo.PipelineBuildTask
 import com.tencent.devops.process.engine.pojo.PipelineModelTask
+import com.tencent.devops.process.engine.service.PipelineBuildDetailService
 import com.tencent.devops.process.engine.service.PipelineRuntimeService
 import com.tencent.devops.process.pojo.PipelineProjectRel
+import com.tencent.devops.process.utils.BK_CI_BUILD_FAIL_TASKNAMES
+import com.tencent.devops.process.utils.BK_CI_BUILD_FAIL_TASKS
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
 import org.springframework.amqp.rabbit.core.RabbitTemplate
@@ -53,6 +58,8 @@ class PipelineTaskService @Autowired constructor(
     val pipelineTaskDao: PipelineTaskDao,
     val pipelineModelTaskDao: PipelineModelTaskDao,
     private val rabbitTemplate: RabbitTemplate,
+    private val pipelineVariableService: BuildVariableService,
+    private val pipelineBuildDetailService: PipelineBuildDetailService,
     private val pipelineRuntimeService: PipelineRuntimeService
 ) {
 
@@ -101,7 +108,8 @@ class PipelineTaskService @Autowired constructor(
             pipelineAtoms?.forEach {
                 val pipelineId = it["pipelineId"] as String
                 val taskParamsStr = it["taskParams"] as? String
-                val taskParams = if (!taskParamsStr.isNullOrBlank()) JsonUtil.getObjectMapper().readValue(taskParamsStr, Map::class.java) as Map<String, Any> else mapOf()
+                val taskParams = if (!taskParamsStr.isNullOrBlank()) JsonUtil.getObjectMapper()
+                    .readValue(taskParamsStr, Map::class.java) as Map<String, Any> else mapOf()
                 if (pipelineAtomVersionInfo.containsKey(pipelineId)) {
                     pipelineAtomVersionInfo[pipelineId]!!.add(taskParams["version"].toString())
                 } else {
@@ -119,7 +127,8 @@ class PipelineTaskService @Autowired constructor(
                     pipelineId = pipelineId,
                     pipelineName = it["pipelineName"] as String,
                     projectCode = it["projectCode"] as String,
-                    atomVersion = pipelineAtomVersionInfo.getOrDefault(pipelineId, mutableListOf<String>()).distinct().joinToString(",")
+                    atomVersion = pipelineAtomVersionInfo.getOrDefault(pipelineId, mutableListOf<String>()).distinct()
+                        .joinToString(",")
                 )
             }
         }
@@ -150,6 +159,101 @@ class PipelineTaskService @Autowired constructor(
     fun removeRetryCache(buildId: String, taskId: String) {
         // 清除该原子内的重试记录
         redisOperation.delete(getRedisKey(buildId, taskId))
+    }
+
+    fun createFailElementVar(buildId: String, projectId: String, pipelineId: String, taskId: String) {
+        logger.info("$buildId| $taskId| atom fail, save var record")
+        val taskRecord = pipelineRuntimeService.getBuildTask(buildId, taskId)
+        val model = pipelineBuildDetailService.getBuildModel(buildId)
+        val failTask = pipelineVariableService.getVariable(buildId, BK_CI_BUILD_FAIL_TASKS)
+        val failTaskNames = pipelineVariableService.getVariable(buildId, BK_CI_BUILD_FAIL_TASKNAMES)
+        var errorElements = ""
+        var errorElementsName = ""
+        if (taskRecord == null) {
+            return
+        }
+        try {
+            val errorElement = findElementMsg(model, taskRecord)
+            errorElements = if (failTask.isNullOrBlank()) {
+                errorElement.first
+            } else {
+                failTask + errorElement.first
+            }
+
+            errorElementsName = if (failTaskNames.isNullOrBlank()) {
+                errorElement.second
+            } else {
+                (failTaskNames + errorElement.second)
+            }
+            logger.info("$buildId| $taskId| atom fail record, tasks:$errorElement, taskNames:$errorElementsName")
+            val valueMap = mutableMapOf<String, Any>()
+            valueMap[BK_CI_BUILD_FAIL_TASKS] = errorElements
+            valueMap[BK_CI_BUILD_FAIL_TASKNAMES] = errorElementsName.substringBeforeLast(",")
+            pipelineVariableService.batchSetVariable(
+                buildId = buildId,
+                projectId = projectId,
+                pipelineId = pipelineId,
+                variables = valueMap
+            )
+        } catch (e: Exception) {
+            logger.warn("createFailElementVar error, msg: $e")
+        }
+    }
+
+    fun removeFailVarWhenSuccess(buildId: String, projectId: String, pipelineId: String, taskId: String) {
+        val failTaskRecord = redisOperation.get(failTaskRedisKey(buildId, taskId))
+        val failTaskNameRecord = redisOperation.get(failTaskNameRedisKey(buildId, taskId))
+        if (failTaskRecord.isNullOrBlank() || failTaskNameRecord.isNullOrBlank()) {
+            logger.info("$buildId|$taskId| retry fail, keep record")
+            return
+        }
+        logger.info("$buildId|$taskId| retry success, start remove fail recode")
+        try {
+            val failTask = pipelineVariableService.getVariable(buildId, BK_CI_BUILD_FAIL_TASKS)
+            val failTaskNames = pipelineVariableService.getVariable(buildId, BK_CI_BUILD_FAIL_TASKNAMES)
+            failTask!!.replace(failTaskRecord!!, "")
+            failTaskNames!!.replace(failTaskNameRecord!!, "")
+            val valueMap = mutableMapOf<String, Any>()
+            valueMap[BK_CI_BUILD_FAIL_TASKS] = failTask
+            valueMap[BK_CI_BUILD_FAIL_TASKNAMES] = failTaskNames
+            pipelineVariableService.batchSetVariable(
+                buildId = buildId,
+                projectId = projectId,
+                pipelineId = pipelineId,
+                variables = valueMap
+            )
+            logger.info("$buildId|$taskId| retry success, success remove fail recode")
+        } catch (e: Exception) {
+            logger.warn("removeFailVarWhenSuccess error, msg: $e")
+        }
+    }
+
+    private fun findElementMsg(model: Model?, taskRecord: PipelineBuildTask): Pair<String, String> {
+        var containerName = ""
+        model?.stages?.forEach { stage ->
+            if (stage.id == taskRecord.stageId) {
+                stage.containers.forEach nextContainer@{ container ->
+                    if (container.id == taskRecord.containerId) {
+                        containerName = container.name
+                        return@forEach
+                    }
+                }
+            }
+        }
+        val failTask = "[${taskRecord.stageId}][$containerName]${taskRecord.taskName} \n"
+        val failTaskName = "${taskRecord.taskName},"
+
+        redisOperation.set(failTaskRedisKey(taskRecord.buildId, taskRecord.taskId), failTask)
+        redisOperation.set(failTaskNameRedisKey(taskRecord.buildId, taskRecord.taskId), failTaskName)
+        return Pair(failTask, failTaskName)
+    }
+
+    private fun failTaskRedisKey(buildId: String, taskId: String): String {
+        return "devops:failTask:redis:key:$buildId:$taskId"
+    }
+
+    private fun failTaskNameRedisKey(buildId: String, taskId: String): String {
+        return "devops:failTaskName:redis:key:$buildId:$taskId"
     }
 
     private fun getRedisKey(buildId: String, taskId: String): String {
